+++ semana12/labeddit/src/router/Router.js
@@ -2,9 +2,9 @@ import React from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom'
 import LoginPage from '../screens/LoginPage/LoginPage'
 import SignUpPage from '../screens/SignUpPage/SignUpPage'
-import FeedPage from '../screens/FeedPage'
+import FeedPage from '../screens/FeedPage/FeedPage'
 import PostDetailsPage from '../screens/PostDetailsPage/PostDetailsPage'
-import ErrorPage from '../screens/ErrorPage'
+import ErrorPage from '../screens/ErrorPage/ErrorPage'
 import Header from '../components/Header/Header';
 
 function Router() {
