--- semana12/labeddit/src/screens/ErrorPage.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function ErrorPage() {
-  return (
-    <div>
-      ErrorPage
-    </div>
-  );
-}
-
-export default ErrorPage;
