+++ semana12/labeddit/src/screens/ErrorPage/styled.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+
+export const MainContainer = styled.div`
+    display: flex;
+    gap: 1em;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    background-color: #DAE0E6;
+    height: 92vh;
+`
