+++ semana12/labeddit/src/screens/FeedPage/styled.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+
+export const PostsContainer = styled.div`
+    display: flex;
+    gap: 1em;
+    flex-direction: column;
+    align-items: center;
+    background-color: #DAE0E6;
+    min-height: 100vh;
+    padding-bottom: 2em;
+`
